fix chat time error； update reward server

+++ app/src/main/java/com/mgc/letobox/happy/me/bean/RewardChatRedpacketBean.java
@@ -22,7 +22,7 @@ public class RewardChatRedpacketBean implements Serializable {
         this.amount = amount;
     }
 
-    public double getChatTime() {
+    public long getChatTime() {
         return chatTime;
     }
